made health check only available on unix
