--- src/health_check/windows.rs
@@ -1,1 +1,0 @@
-const HEALTH_CHECK_PATH: &str = r"\\.\pipe\wisdom\swat-collector\health";
